Separate DigitalSignal and DigitalSequence (draft)
